feat(#S11P21D109-270): 경매 등록 api 작성

+++ android/domain/src/main/java/com/semonemo/domain/model/AuctionRegisterRequest.kt
@@ -2,5 +2,5 @@ package com.semonemo.domain.model
 
 data class AuctionRegisterRequest(
     val nftId: Long = 0L,
-    val startPrice: Int = 0,
+    val startPrice: Long = 0,
 )
